+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nextui-org/popover
 
+## 0.0.0-dev-v2-20230409142211
+
+### Patch Changes
+
+- Left/Right labels changed to Start/End, tests fixed
+- Updated dependencies
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230409142211
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230409142211
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230409142211
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230409142211
+  - @nextui-org/button@0.0.0-dev-v2-20230409142211
+  - @nextui-org/system@0.0.0-dev-v2-20230409142211
+  - @nextui-org/theme@0.0.0-dev-v2-20230409142211
+
 ## 0.0.0-dev-v2-20230409022144
 
 ### Patch Changes