+++ packages/core/react/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @nextui-org/react
 
+## 0.0.0-dev-v2-20230409142211
+
+### Patch Changes
+
+- Left/Right labels changed to Start/End, tests fixed
+- Updated dependencies
+  - @nextui-org/pagination@0.0.0-dev-v2-20230409142211
+  - @nextui-org/accordion@0.0.0-dev-v2-20230409142211
+  - @nextui-org/checkbox@0.0.0-dev-v2-20230409142211
+  - @nextui-org/dropdown@0.0.0-dev-v2-20230409142211
+  - @nextui-org/progress@0.0.0-dev-v2-20230409142211
+  - @nextui-org/popover@0.0.0-dev-v2-20230409142211
+  - @nextui-org/snippet@0.0.0-dev-v2-20230409142211
+  - @nextui-org/spinner@0.0.0-dev-v2-20230409142211
+  - @nextui-org/tooltip@0.0.0-dev-v2-20230409142211
+  - @nextui-org/avatar@0.0.0-dev-v2-20230409142211
+  - @nextui-org/button@0.0.0-dev-v2-20230409142211
+  - @nextui-org/switch@0.0.0-dev-v2-20230409142211
+  - @nextui-org/badge@0.0.0-dev-v2-20230409142211
+  - @nextui-org/input@0.0.0-dev-v2-20230409142211
+  - @nextui-org/radio@0.0.0-dev-v2-20230409142211
+  - @nextui-org/card@0.0.0-dev-v2-20230409142211
+  - @nextui-org/chip@0.0.0-dev-v2-20230409142211
+  - @nextui-org/code@0.0.0-dev-v2-20230409142211
+  - @nextui-org/drip@0.0.0-dev-v2-20230409142211
+  - @nextui-org/link@0.0.0-dev-v2-20230409142211
+  - @nextui-org/user@0.0.0-dev-v2-20230409142211
+  - @nextui-org/system@0.0.0-dev-v2-20230409142211
+  - @nextui-org/theme@0.0.0-dev-v2-20230409142211
+
 ## 0.0.0-dev-v2-20230409022144
 
 ### Patch Changes